+++ veterans/tsconfig.json
@@ -8,7 +8,8 @@
       "@lib/*": ["lib/*"],
       "@keystone/*": ["keystone/*"],
       "hooks/*": ["hooks/*"],
-      "constants/*": ["constants/*"]
+      "constants/*": ["constants/*"],
+      "utils/*": ["utils/*"]
     },
     "jsx": "preserve",
     "target": "esnext",
